* Added cast to longtype when reading timestamp column from kafka dataframe. * Modified spark context management in tests environment

+++ src/main/scala/com/org/challenge/stream/jobs/kafka/PageViewsStream.scala
@@ -9,9 +9,8 @@ import com.org.challenge.stream.schemas.SchemaManagement
 import com.org.challenge.stream.transformation.BaseTransform
 import com.org.challenge.stream.utils.Utils
 import com.org.challenge.stream.writers.BaseWriter
-import org.apache.commons.math3.transform.TransformType
 import org.apache.spark.sql.functions.{col, from_json, from_unixtime}
-import org.apache.spark.sql.types.{StringType, StructType, TimestampType}
+import org.apache.spark.sql.types.{LongType, StringType, StructType, TimestampType}
 import org.apache.spark.sql.{DataFrame, SparkSession}
 
 sealed trait SchemaType {
@@ -137,7 +136,7 @@ class PageViewsStream(spark: SparkSession, params: Params) extends StreamJob[Par
           this.log.info(s"Watermark column is ${watermarkCol}")
           this.log.info(s"Maximum delay for messages is ${this.topicsDelayPair.get(t).get} seconds")
           df
-            .select(from_json(col("value").cast(StringType), schemaForTopic).as("data"), col(watermarkCol))
+            .select(from_json(col("value").cast(StringType), schemaForTopic).as("data"), col(watermarkCol).cast(LongType))
             .select(col("data.*"), from_unixtime(col(watermarkCol)).cast(TimestampType).as(watermarkCol))
             .withWatermark(watermarkCol, s"${this.topicsDelayPair.get(t).get} seconds")
         }
